requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
constantly==15.1.0
cryptography==37.0.2
cssselect2==0.6.0
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.5
django-baton==2.3.0
django-ckeditor==6.4.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-grappelli==3.0.3
django-hosts==5.1
django-import-export==2.8.0
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-material-admin==1.8.6
django-multiselectfield==0.1.12
django-widget-tweaks==1.4.8
environ==1.0
et-xmlfile==1.1.0
filelock==3.7.0
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
lxml==4.8.0
MarkupPy==1.14
mysql-connector==2.2.9
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
Pillow==9.1.0
platformdirs==2.5.2
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
pyOpenSSL==22.0.0
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
razorpay==1.3.0
reportlab==3.6.9
requests==2.27.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tablib==3.2.1
tinycss2==1.1.1
tqdm==4.64.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
virtualenv==20.0.31
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.7
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
